Standardize TFRecord JPG/PNG generation between cuCIM & Libvips
- This also fixes Libvips OME-TIFF reading, which was previously stuck reading images in monochrome format due to a libvips bug
